requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.8
django-cities-light==3.9.2
progressbar2==4.2.0
psycopg2-binary==2.9.5
python-utils==3.5.2
pytz==2022.7.1
sqlparse==0.4.3
Unidecode==1.3.6
